# Speedup of `estimate_yj()`