--- a/R/YeoJohnson.R
+++ b/R/YeoJohnson.R
@@ -150,7 +150,7 @@ print.step_YeoJohnson <-
 #' @export
 #' @keywords internal
 #' @rdname recipes-internal
-yj_transform <- function(x, lambda, eps = .001) {
+yj_transform <- function(x, lambda, ind_neg = NULL, eps = 0.001) {
   if (is.na(lambda))
     return(x)
   if (!inherits(x, "tbl_df") || is.data.frame(x)) {
@@ -160,8 +160,12 @@ yj_transform <- function(x, lambda, eps = .001) {
       x <- as.vector(x)
   }
 
-  not_neg <- which(x >= 0)
-  is_neg <- which(x < 0)
+  if (is.null(ind_neg)) {
+    dat_neg <- x < 0
+    ind_neg <- list(is = which(dat_neg), not = which(!dat_neg))
+  }
+  not_neg <- ind_neg[["not"]]
+  is_neg <- ind_neg[["is"]]
 
   nn_trans <- function(x, lambda)
     if (abs(lambda) < eps)
@@ -187,31 +191,25 @@ yj_transform <- function(x, lambda, eps = .001) {
 ## Helper for the log-likelihood calc for eq 3.1 of Yeo, I. K.,
 ## & Johnson, R. A. (2000). A new family of power transformations
 ## to improve normality or symmetry. Biometrika. page 957
-
-ll_yj <- function(lambda, y, eps = .001) {
-  y <- y[!is.na(y)]
+ll_yj <- function(lambda, y, ind_neg, const, eps = 0.001) {
   n <- length(y)
-  nonneg <- all(y > 0)
-  y_t <- yj_transform(y, lambda)
+  y_t <- yj_transform(y, lambda, ind_neg)
   mu_t <- mean(y_t)
   var_t <- var(y_t) * (n - 1) / n
-  const <- sum(sign(y) * log(abs(y) + 1))
   res <- -.5 * n * log(var_t) + (lambda - 1) * const
   res
 }
 
 ## eliminates missing data and returns -llh
-yj_obj <- function(lam, dat){
-  dat <- dat[complete.cases(dat)]
-  ll_yj(lambda = lam, y = dat)
+yj_obj <- function(lam, dat, ind_neg, const) {
+  ll_yj(lambda = lam, y = dat, ind_neg = ind_neg, const = const)
 }
 
 ## estimates the values
 #' @export
 #' @keywords internal
 #' @rdname recipes-internal
-estimate_yj <- function(dat, limits = c(-5, 5), num_unique = 5,
-                        na_rm = TRUE) {
+estimate_yj <- function(dat, limits = c(-5, 5), num_unique = 5, na_rm = TRUE) {
   na_rows <- which(is.na(dat))
   if (length(na_rows) > 0) {
     if (na_rm) {
@@ -224,11 +222,18 @@ estimate_yj <- function(dat, limits = c(-5, 5), num_unique = 5,
   eps <- .001
   if (length(unique(dat)) < num_unique)
     return(NA)
+  dat_neg <- dat < 0
+  ind_neg <- list(is = which(dat_neg), not = which(!dat_neg))
+
+  const <- sum(sign(dat) * log(abs(dat) + 1))
+
   res <- optimize(
     yj_obj,
     interval = limits,
     maximum = TRUE,
     dat = dat,
+    ind_neg = ind_neg,
+    const = const,
     tol = .0001
   )
   lam <- res$maximum