@juliasilge 2021-09-20
We probably want to get this update into the `NEWS` @topepo